requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
sqlalchemy
asyncpg
aiosqlite
pytest
pytest-asyncio
alembic